Fixed broken logic, cleaned up file heirarchy , higher resolution images
The carousel was no longer working. Remade a new one using slider.js using the same timings and style. Did remove the hover effect as it was redundant in my opinion. Also re-sized all the images to make them higher resolution.All the files used to be under 1 direcotry with .html but I moved them all to directories and changed them to index.html. Makes everything look cleaner and makes the url cleaner as well.